python-ai-services/requirements.txt
# PydanticAI Integration Requirements
pydantic-ai[openai,anthropic,gemini]==0.0.14
pydantic==2.10.4
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-cloud-aiplatform==1.78.0
google-cloud-pubsub==2.30.1
google-cloud-functions==1.19.2
google-cloud-firestore==2.20.0
redis==5.2.1
asyncio-mqtt==0.16.2
websockets==14.1
loguru==0.7.3
python-multipart==0.0.20
httpx==0.28.1
orjson==3.10.12
structlog==24.4.0
psutil==6.1.0

# Trading specific
numpy==2.2.1
pandas==2.2.3
scipy==1.14.1
scikit-learn==1.6.0
ta-lib==0.4.32
yfinance==0.2.50
alpha-vantage==2.3.1

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
isort==5.13.2
mypy==1.13.0
supabase-py
python-dotenv
requests
ag-ui-protocol
crewai>=0.28.0
crewai-tools>=0.1.0
datamodel-code-generator>=0.25.2
langchain-google-genai>=0.1.0

# MemGPT for agent memory
pymemgpt[postgres,pgvector]~=0.3.5 
# If the above doesn't pull these, they might be needed explicitly:
# psycopg2-binary # For PostgreSQL connection
# pgvector # For vector similarity search in PostgreSQL

# For SharedKnowledgeService embedding generation
sentence-transformers>=2.2.0